+++ src/main/java/lotto/domain/Lotto.java
@@ -14,6 +14,14 @@ public class Lotto {
         this.lottoNumbers = lottoNumbers;
     }
 
+    public int countMatchedNumber(Lotto answerLotto) {
+        List<Integer> answerLottoNumbers = answerLotto.getLottoNumbers();
+
+        return (int) lottoNumbers.stream()
+                .filter(answerLottoNumbers::contains)
+                .count();
+    }
+
     public List<Integer> getLottoNumbers() {
         return lottoNumbers;
     }
