+++ src/test/java/lotto/domain/LottoTest.java
@@ -26,4 +26,12 @@ public class LottoTest {
         ).isInstanceOf(IllegalArgumentException.class)
                 .hasMessageContaining("[ERROR]");
     }
+
+    @Test
+    @DisplayName("로또번호 정답 갯수를 확인한다.")
+    void 로또번호_정답_갯수_확인() {
+        Lotto lotto = new Lotto(Arrays.asList(1, 2, 3, 4, 5, 6));
+        Lotto answerLotto = new Lotto(Arrays.asList(4, 5, 6, 7, 8, 9));
+        assertThat(lotto.countMatchedNumber(answerLotto)).isEqualTo(3);
+    }
 }
